GameScene

+++ Assets/_Game/Scripts/GameScene/Enemies/EnemyInfo.cs
@@ -5,6 +5,6 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Enemy", menuName = "Scriptable/Enemy")]
 public class EnemyInfo : ScriptableObject
 {
-    [field: SerializeField] public int Speed { get; private set; }
-    [field: SerializeField] public int Health { get; private set; }
+    [field: SerializeField] public float Speed { get; private set; }
+    [field: SerializeField] public float Health { get; private set; }
 }
